pyproject.toml
[tool.poetry]
name = "gitlearning"
version = "0.1.0"
description = ""
authors = ["DanielAndDastan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
drf-yasg = "^1.21.4"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
